LayoutTests/fast/dom/navigator-vibration.html

<html>
<head>
<script src="../js/resources/js-test-pre.js"></script>
</head>
<body>
<script>
description('Test for the Vibration API.');

shouldBe("navigator.vibrate(0);", "undefined");
shouldBe("navigator.vibrate([]);", "undefined");
shouldBe("navigator.vibrate(1000);", "undefined");
shouldBe("navigator.vibrate([1000, 300, 500]);", "undefined");
shouldBe("navigator.vibrate(4294967295);", "undefined");
shouldThrow("navigator.vibrate();");
</script>
<script src="../js/resources/js-test-post.js"></script>
</body>
</html>
